QDisk: corner case fix
this commit fixing a corner case.
If we write data to disk with exactly of `disk_buf_size`,
then write head is "wrap" and we write more data to disk.
A read head will read to the end of the file, BUT isn't "wrapping"
and try to read after the file, that's the bug.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>
Signed-off-by: Balázs Barkó <Balazs.Barko@oneidentity.com>